IB/cm: Clear all associated AV's ports when remove a cm device
When remove a cm device all ports are removed as well, so all AV's ports
need to be cleared.
This patch adds a cm_id_priv list for each cm_device; For a cm_id when
initializing it's primary AV it is added to this list, so when
removing the device all cm_id's on this list will be removed from this
list and have its av->port and alt_av->port pointer cleared.

Signed-off-by: Mark Zhang <markzhang@nvidia.com>
Signed-off-by: Leon Romanovsky <leonro@nvidia.com>