Revert "bpf: Remove unnecessary assertion on fp_old"
This reverts commit 5bf2fc1f9c88397b125d5ec5f65b1ed9300ba59d.

Commits from @umn.edu addresses have been found to be submitted in "bad
faith" to try to test the kernel community's ability to review "known
malicious" changes.  The result of these submissions can be found in a
paper published at the 42nd IEEE Symposium on Security and Privacy
entitled, "Open Source Insecurity: Stealthily Introducing
Vulnerabilities via Hypocrite Commits" written by Qiushi Wu (University
of Minnesota) and Kangjie Lu (University of Minnesota).

Because of this, all submissions from this group must be reverted from
the kernel tree and will need to be re-reviewed again to determine if
they actually are a valid fix.  Until that work is complete, remove this
change to ensure that no problems are being introduced into the
codebase.

Cc: Aditya Pakki <pakki001@umn.edu>
Cc: Daniel Borkmann <daniel@iogearbox.net>
Cc: https
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>