Revert "drm/gma500: fix memory disclosures due to uninitialized bytes"
This reverts commit ec3b7b6eb8c90b52f61adff11b6db7a8db34de19.

Commits from @umn.edu addresses have been found to be submitted in "bad
faith" to try to test the kernel community's ability to review "known
malicious" changes.  The result of these submissions can be found in a
paper published at the 42nd IEEE Symposium on Security and Privacy
entitled, "Open Source Insecurity: Stealthily Introducing
Vulnerabilities via Hypocrite Commits" written by Qiushi Wu (University
of Minnesota) and Kangjie Lu (University of Minnesota).

Because of this, all submissions from this group must be reverted from
the kernel tree and will need to be re-reviewed again to determine if
they actually are a valid fix.  Until that work is complete, remove this
change to ensure that no problems are being introduced into the
codebase.

Cc: Kangjie Lu <kjlu@umn.edu>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Cc: https
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>