fs/ext4: remove redundant initialization of variable busy
The variable status is being initialized with a value that is never
read and it is being updated later with a new value. The initialization
is redundant and could be removed. Also put the variable declarations
into reverse christmas tree order. Finally, we add the log printing and
ext4_mb_show_pa() for troubleshooting, they are enabled only when
CONFIG_EXT4_DEBUG is set.

Signed-off-by: Wen Yang <wenyang@linux.alibaba.com>
Cc: "Theodore Ts'o" <tytso@mit.edu>
Cc: Andreas Dilger <adilger.kernel@dilger.ca>
Cc: Baoyou Xie <baoyou.xie@alibaba-inc.com>
Cc: linux-ext4@vger.kernel.org
Cc: linux-kernel@vger.kernel.org