lpfc: Fix bad memory access during VPD DUMP mailbox command
The dump command for reading a region passes a requested read length
specified in words (4byte units). The response overwrites the same
field with the actual number of bytes read.

The mailbox handler for DUMP which reads VPD data (region 23) is
treating the response field as if it were still a word_cnt, thus
multiplying it by 4 to set the read's "length". Given the read value
was calculated based on the size of the read buffer, the longer
response length runs off the end of the buffer.

Fix by reworking the code to use the response field as a byte count.

Co-developed-by: Justin Tee <justin.tee@broadcom.com>
Signed-off-by: Justin Tee <justin.tee@broadcom.com>
Signed-off-by: James Smart <jsmart2021@gmail.com>

Notes:
	ECD submit header:

	*******************

	COMMIT-TITLE: Back out svn rev 73205 - causing KASAN error
	COMMIT: 78630
	COMMIT-BZ: 244614
	COMMIT-QC:

  r78630 | jinfante | 2021-04-05 12:40:48 -0700 (Mon, 05 Apr 2021) | 25 lines

  Back out svn rev 73205 - causing KASAN error

  Bugs: 244614

  Reviewers: rkennedy pely

  Symptoms:
  KASAN detected issues
  page dumped because: kasan: bad access detected

  Cause:
  svn rev 73205

  Fix:
  dump mbox takes word count as input, byte count as output.

  Notes:

  Unit Testing:
  Added log message on VPD dump to verify output

  System Testing:

  Review_ID: https://cmengapps1.lvn.broadcom.net/r/34471/