leds: trigger/tty: feature data direction
The current implementation just signals a visible feedback on all kind of
activity on the corresponding TTY. But sometimes it is useful to see what
kind of activity just happens. This change adds the capability to filter
the direction of TTY's data flow. It enables a user to forward both
directions to separate LEDs for tx and rx on demand. Default behavior is
still both directions.

Signed-off-by: Juergen Borleis <jbe@pengutronix.de>