kfifo: fix ternary sign extension bugs
The intent with this code was to return negative error codes but instead
it returns positives.

The problem is how type promotion works with ternary operations.  These
functions return long, "ret" is an int and "copied" is a u32.  The
negative error code is first cast to u32 so it becomes a high positive
and then cast to long where it's still a positive.

We could fix this by declaring "ret" as a ssize_t but let's just get
rid of the ternaries instead.

Fixes: 5bf2b19320ec ("kfifo: add example files to the kernel sample directory")
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>