f2fs: set prefree as free segments after clear prefree segments
For now, when do_checkpoint fails, the prefree bitmap is not cleared,
but these segments are already in the free state. If these segments
are used, the segments in use will be reset to the free state when
f2fs_clear_prefree_segments is called next time.

So move set_prefree_as_free_segments after clear_prefree_segments.

Signed-off-by: Yangtao Li <frank.li@vivo.com>